WA-238 Refactor transaction tests
